+++ packages/core/src/components/viewer-grid/components/main-panel.tsx
@@ -1,39 +1,31 @@
 import { EventEmitter, FunctionalComponent, h } from "@stencil/core";
 import clsx from "clsx";
 
-import { ViewerGridChangeSizeAnimationEndEvent, ViewerGridMode, ViewerGridPanelSize } from "../viewer-grid.interfaces";
-
-import { SizingButtons } from "./sizing-buttons";
+import { ViewerGridChangeSizeAnimationEndEvent, ViewerGridPanelSize } from "../viewer-grid.interfaces";
 
 export interface ViewerGridMainPanelProps {
-  mode: ViewerGridMode;
   tabView: boolean;
   mainSize: ViewerGridPanelSize;
   documentPanelOpen: boolean;
   mainPanelExpanded: boolean;
   mainPanelHidden: boolean;
-  shrinkMain: () => void;
-  expandMain: () => void;
   toggleMainPanel: () => void;
   dsoMainSizeChangeAnimationEnd: EventEmitter<ViewerGridChangeSizeAnimationEndEvent>;
 }
 
 export const MainPanel: FunctionalComponent<ViewerGridMainPanelProps> = ({
-  mode,
   tabView,
   mainSize,
   documentPanelOpen,
   mainPanelExpanded,
   mainPanelHidden,
-  shrinkMain,
-  expandMain,
   toggleMainPanel,
   dsoMainSizeChangeAnimationEnd,
 }) => (
   <div
     class={clsx("dso-main-panel", {
       compact: !tabView && documentPanelOpen && !mainPanelExpanded,
-      contracted: !tabView && !documentPanelOpen && !mainPanelExpanded && mode === "vdk",
+      contracted: !tabView && !documentPanelOpen && !mainPanelExpanded,
       expanded: !tabView && documentPanelOpen && mainPanelExpanded,
       collapsed: mainPanelHidden,
     })}
@@ -43,24 +35,14 @@ export const MainPanel: FunctionalComponent<ViewerGridMainPanelProps> = ({
       }
     }}
   >
-    {!tabView &&
-      ((mode === "vrk" && (
-        <SizingButtons
-          panelLabel="Hoofdpaneel"
-          size={mainSize}
-          expand={expandMain}
-          shrink={shrinkMain}
-          placement="left"
-        />
-      )) ||
-        (mode === "vdk" && documentPanelOpen && (
-          <div class="toggle-button">
-            <button type="button" onClick={toggleMainPanel}>
-              <span class="sr-only">Zoeken paneel {mainPanelHidden ? "tonen" : "verbergen"}</span>
-              <dso-icon icon={mainPanelHidden ? "chevron-right" : "chevron-left"}></dso-icon>
-            </button>
-          </div>
-        )))}
+    {!tabView && documentPanelOpen && (
+      <div class="toggle-button">
+        <button type="button" onClick={toggleMainPanel}>
+          <span class="sr-only">Zoeken paneel {mainPanelHidden ? "tonen" : "verbergen"}</span>
+          <dso-icon icon={mainPanelHidden ? "chevron-right" : "chevron-left"}></dso-icon>
+        </button>
+      </div>
+    )}
     <div class={clsx("content", { invisible: mainPanelHidden })}>
       <slot name="main" />
     </div>