+++ packages/core/src/components/viewer-grid/viewer-grid.interfaces.ts
@@ -15,14 +15,6 @@ export interface ViewerGridCloseFilterpanelEvent {
   originalEvent: MouseEvent | Event;
 }
 
-export interface ViewerGridFilterpanelCancelEvent {
-  originalEvent: MouseEvent | Event;
-}
-
-export interface ViewerGridFilterpanelApplyEvent {
-  originalEvent: MouseEvent | Event;
-}
-
 export interface ViewerGridActiveTabSwitchEvent {
   tab: ViewerGridTab;
 }
@@ -35,24 +27,16 @@ export interface ViewerGridMainToggleEvent {
   hide: boolean;
 }
 
-export type ViewerGridMode = "vdk" | "vrk";
-
 export type ViewerGridPanelSize = "small" | "medium" | "large";
 
 export type ViewerGridLabelSizeMap = { [key in ViewerGridPanelSize]: string };
 
-export const viewerGridVrkTabs = ["main", "map"] as const;
-export type ViewerGridVrkTab = (typeof viewerGridVrkTabs)[number];
-
-export const viewerGridVdkTabs = ["search", "map", "document"] as const;
-export type ViewerGridVdkTab = (typeof viewerGridVdkTabs)[number];
-
-export type ViewerGridTab = ViewerGridVrkTab | ViewerGridVdkTab;
+export const viewerGridTabs = ["search", "map", "document"] as const;
+export type ViewerGridTab = (typeof viewerGridTabs)[number];
 
 export type ViewerGridTabLabelMap = { [key in ViewerGridTab]: string };
 
 export const viewerGridTabLabelMap: ViewerGridTabLabelMap = {
-  main: "Hoofdpaneel",
   map: "Kaart",
   document: "Document",
   search: "Zoeken",